docs/requirements.txt
#installs any requirements in a virtual environment, which is then where sphinx-apidoc is run
#needed when running autodoc
sphinx==7.1.2
sphinx-rtd-theme==1.3.0rc1
sphinxcontrib.youtube
myst_parser